refactor params helper
